Make sure extension metadata properties are not including timestamps
(cherry picked from commit e14daa1f267d8d578eb5f5bfd8fe907e16b826ca)
